:recycle: (aula 5) ajuste incluindo menu nos exercicios da aula 5
