+++ components/footer.php
@@ -2,7 +2,9 @@
 <footer class="bg-orange-500 text-white">
   <div class="container mx-auto text-center py-8">
     <p class="text-sm mb-1 font-bold uppercase">João Luiz Vicente Junior &copy; 2025. Todos os direitos reservados.</p>
-    <p class="text-xs text-orange-200">Desenvolvimento Web 2 - Gestão de Tecnologia - IFPR</p>
+    <p class="text-xs text-orange-200">
+      <a href="/" class="hover:underline">Desenvolvimento Web 2 - Gestão de Tecnologia - IFPR - Campus Pinhais</a>
+    </p>
   </div>
 </footer>
 
@@ -20,6 +22,10 @@
     });
   });
 </script>
+
+<?php if (isset($configuration['hasBootstrap'])): ?>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe" crossorigin="anonymous"></script>
+<?php endif; ?>
 
 </body>
